11.15合并

See merge request java/SuperCode3.0/platform-marketing-supercode-3.0!18

+++ src/main/java/com/jgw/supercodeplatform/prizewheels/domain/model/WheelsReward.java
@@ -68,11 +68,11 @@ public class WheelsReward implements Serializable {
     private Integer initialStock;
 
 
-    private Integer randLowMoney;
+    private Double randLowMoney;
 
-    private Integer randHighMoney;
+    private Double randHighMoney;
 
-    private Integer fixedMoney;
+    private Double fixedMoney;
 
     private Integer moneyType;
